requirements.txt
cachetools==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
datatable==1.1.0
et-xmlfile==1.1.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
molgenis-emx2-pyclient==10.48.9
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.3
pandas==2.1.4
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
randomhash==0.6.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
xlsx2csv==0.8.2
XlsxWriter==3.2.0
xxhash==3.4.1
